wE CAN Talk to teh Serv

+++ Kluster/src/com/cs446/kluster/User.java
@@ -1,14 +1,16 @@
 package com.cs446.kluster;
 
+import java.math.BigInteger;
+
 import android.accounts.Account;
 
 
 public class User {
-	private int mUserID;
+	private BigInteger mUserID;
 	private Account mAccount;
 	/**TODO: Add as needed */
 	
-	public int getID() {
+	public BigInteger getID() {
 		return mUserID;
 	}
 	
@@ -16,7 +18,7 @@ public class User {
 		return mAccount;
 	}
 	
-	public User(int id, Account account) {
+	public User(BigInteger id, Account account) {
 		mUserID = id;
 		mAccount = account;
 	}
